requirements.txt
requests
facebook-scraper
configparser~=5.0.2
pytz~=2021.1
python-telegram-bot